requirements.txt
argparse==1.2.1
selenium==2.44.0
splinter==0.6.0
wheezy.core==0.1.136
wsgiref==0.1.2
